+++ task06_momento_memory/index.html
@@ -26,5 +26,6 @@
   </form>
   <script type="module" src="./script/index.js"></script>
   <script type="module" src="./script/display.js"></script>
+  <script type="module" src="./script/permutation.js"></script>
 </body>
 </html>
