+++ task06_momento_memory/script/index.js
@@ -1,4 +1,5 @@
 import { display } from "./display.js";
+import { Permutation } from "./permutation.js";
 
 const btn = document.getElementById('btnSubmit');
 btn. addEventListener('click', work);
@@ -7,25 +8,11 @@ function work() {
   display("clear");
   let word = document.getElementById("word").value;
   clearInpits();
-  display(permutation(word));
+  const permutation = new Permutation(word);
+  display(permutation.getResult());
+  setTimeout(display,5000,"clear");
 }
 
 function clearInpits() {
   document.getElementById('word').value = '';
-}
-
-
-function permutation(word) {
-  let arr = [word];
-  let anagrams = {};
-  arr.forEach(function (str) {
-    let recurse = function (ana, str) {
-      if (str === '')
-        anagrams[ana] = 1;
-      for (let i = 0; i < str.length; i++)
-        recurse(ana + str[i], str.slice(0, i) + str.slice(i + 1));
-    };
-    recurse('', str);
-  });
-  display(Object.keys(anagrams));
 }
